# Improve ChunkedCollector performance slightly