# Summary

This change improves performance of `ChunkedCollector` slightly, but also highlights an important issue (which I strongly believe will improve in Go 1.5, hopefully I will have time to test with 1.5 soon to verify). For now see the workaround mentioned in [Solution](#Solution) section 2.

(also see relevant tests and data at [Appdash PR #80](https://github.com/sourcegraph/appdash/pull/80))

### Presumptions

I believe that `ChunkedCollector` when placing items into it's underlying map (which holds onto them for < 500ms in our prod. apps, and just 10ms in our tests) is indirectly causing 20-41k tiny heap objects to be allocated (but the Go code is NOT leaking them) and that the Go runtime is poorly equipped in Go 1.4 to either: 

1. Reuse these small objects behind the map so that many allocations do not need to occur.
2. Release the memory allocated by these small objects back to the OS (seen as `top RES` above).

### History

I have also observed results similiar to this before in another project of mine (which unlike the case here, actually allocated several hundred thousand small objects per second), leaving me with ~12ms GC times on average. I believe these issues to be one in the same and likely fixed or improved significantly with the Go 1.5 GC improvements.

### Analysis:

- [Tests 3](https://github.com/sourcegraph/appdash/pull/80#issuecomment-128191959) [and 4](https://github.com/sourcegraph/appdash/pull/80#issuecomment-128192006) show that garbage collection does bring the number of heap-objects back down to a reasonable amount, but still worse than it should be (none of those objects are in use!).
- [Test 1](https://github.com/sourcegraph/appdash/pull/80#issuecomment-128191822) shows that invoking both `GC` and `FreeOSMemory` brings the number of heap-objects back to the lowest numbers (i.e. what they should be). 
- Only [test 1](https://github.com/sourcegraph/appdash/pull/80#issuecomment-128191822) (which invokes `FreeOSMemory`) appears to release any memory back to the host OS (Note: it is likely that Go invokes this function periodically and our test just doesn't run long enough to hit this, but I am not certain).

### Solution:

In order to fix this issue i have looked at two things as a solution:

1. Attempt to lower the amount of heap objects that Appdash `ChunkedCollector` must allocate.
  - This is not possible, because (after PR #80 is merged) it is literally just a very simple map and is being cleared to nil upon each `ChunkedCollector.Flush`.
  - Using anything other than a map (e.g. a slice) is probably not viable because of the number of spans that will end up in the map, we would effectively be trading memory usage for CPU usage. Not good.
  - However, I was able to identify and remove (in PR #80) a single slice that was uneeded, this cuts the total allocations nearly in half.
2. Setup a background goroutine that occasionally invokes `runtime.GC` twice (needed to trigger a full GC) followed by one `FreeOSMemory` call.
  - Go by default triggers a GC only after the heap has grown 100%, which is normally fine, but in this case it appears to be unable to recollect these small objects and reuse them, so the heap grows nearly unbounded in size.
  - It is likely this issue would appear in apps today as a unbounded heap growth up to entire memory consumption, followed by either a (very) long GC or an OOM runtime panic.
  - A good value for which people may call these three functions in sequence might be every 5-30 minutes, just enough so that these small allocations are returned to the host OS (and can be reqaquired by Go subsequently) before the heap grows too large, but not such a quick value that it would hinder performance.

Example:

```
func init() {
	// Workaround until Go 1.5 is released, as described at:
	//
	//  https://github.com/sourcegraph/appdash/pull/80
	//
	go func() {
		for{
			time.Sleep(5 * time.Minute)
			runtime.GC()
			runtime.GC()
			debug.FreeOSMemory()
		}
	}()
}
```