@slimsag 2015-08-05
CI failure is unrelated to this change -- I've also tested locally and it passes all tests fine.