@slimsag 2015-08-06
# Test 1

| Test Feature | Enabled |
|--------------|---------|
| 10 explicit runtime.GC calls? | Yes |
| 10 explicit runtime/debug.FreeOSMemory calls? | Yes |

```
$ go test -tags=danger -memprofilerate=1 -memprofile=mem.out -bench=BenchmarkChunkedCollector1mil -run=NONE -v
PASS
BenchmarkChunkedCollector1mil	

[pre] 1000000-collections:
  General statistics
    - Alloc       82KB   (allocated and still in use)                  
    - TotalAlloc  863KB  (allocated (even if freed))                   
    - Sys         2MB    (obtained from system (sum of XxxSys below))  
    - Lookups     5      (number of pointer lookups)                   
    - Mallocs     15976  (number of mallocs)                           
    - Frees       15446  (number of frees)                             

  Main allocation heap statistics
    - HeapAlloc     82KB   (allocated and still in use)         
    - HeapSys       848KB  (obtained from system)               
    - HeapIdle      464KB  (in idle spans)                      
    - HeapInuse     384KB  (in non-idle span)                   
    - HeapReleased  0B     (released to the OS)                 
    - HeapObjects   530    (total number of allocated objects)  

  Low-level fixed-size structure allocator statistics
    - StackInuse   176KB  (used by stack allocator right now)                   
    - StackSys     176KB  (obtained from system)                                
    - MSpanInuse   4KB    (mspan structures / in use now)                       
    - MSpanSys     16KB   (obtained from system)                                
    - MCacheInuse  1KB    (in use now))                                         
    - MCacheSys    16KB   (mcache structures / obtained from system)            
    - BuckHashSys  1MB    (profiling bucket hash table / obtained from system)  
    - GCSys        70KB   (GC metadata / obtained form system)                  
    - OtherSys     269KB  (other system allocations)                            

  

[post] 1000000-collections:
    General statistics
      - Alloc       163MB  (allocated and still in use)                  
      - TotalAlloc  314MB  (allocated (even if freed))                   
      - Sys         178MB  (obtained from system (sum of XxxSys below))  
      - Lookups     5      (number of pointer lookups)                   
      - Mallocs     75075  (number of mallocs)                           
      - Frees       45342  (number of frees)                             

    Main allocation heap statistics
      - HeapAlloc     163MB  (allocated and still in use)         
      - HeapSys       165MB  (obtained from system)               
      - HeapIdle      1MB    (in idle spans)                      
      - HeapInuse     164MB  (in non-idle span)                   
      - HeapReleased  0B     (released to the OS)                 
      - HeapObjects   29733  (total number of allocated objects)  

    Low-level fixed-size structure allocator statistics
      - StackInuse   176KB  (used by stack allocator right now)                   
      - StackSys     176KB  (obtained from system)                                
      - MSpanInuse   165KB  (mspan structures / in use now)                       
      - MSpanSys     176KB  (obtained from system)                                
      - MCacheInuse  1KB    (in use now))                                         
      - MCacheSys    16KB   (mcache structures / obtained from system)            
      - BuckHashSys  1MB    (profiling bucket hash table / obtained from system)  
      - GCSys        10MB   (GC metadata / obtained form system)                  
      - OtherSys     1MB    (other system allocations)                            

[garbage collections]
  0. GC - 12.675495ms
  1. GC - 9.219848ms
  2. GC - 8.557734ms
  3. GC - 8.729595ms
  4. GC - 8.574088ms
  5. GC - 8.921975ms
  6. GC - 8.551715ms
  7. GC - 8.605256ms
  8. GC - 8.704241ms
  9. GC - 990.798µs

[free os memory]
  0. FreeOSMemory - 2.886292ms
  1. FreeOSMemory - 271.009µs
  2. FreeOSMemory - 228.818µs
  3. FreeOSMemory - 218.233µs
  4. FreeOSMemory - 214.515µs
  5. FreeOSMemory - 213.408µs
  6. FreeOSMemory - 213.43µs
  7. FreeOSMemory - 222.377µs
  8. FreeOSMemory - 222.013µs
  9. FreeOSMemory - 215.682µs

    

[after GC & FreeOSMemory] 1000000-collections:
      General statistics
        - Alloc       81KB   (allocated and still in use)                  
        - TotalAlloc  314MB  (allocated (even if freed))                   
        - Sys         178MB  (obtained from system (sum of XxxSys below))  
        - Lookups     5      (number of pointer lookups)                   
        - Mallocs     76490  (number of mallocs)                           
        - Frees       75942  (number of frees)                             

      Main allocation heap statistics
        - HeapAlloc     81KB   (allocated and still in use)         
        - HeapSys       165MB  (obtained from system)               
        - HeapIdle      165MB  (in idle spans)                      
        - HeapInuse     376KB  (in non-idle span)                   
        - HeapReleased  165MB  (released to the OS)                 
        - HeapObjects   548    (total number of allocated objects)  

      Low-level fixed-size structure allocator statistics
        - StackInuse   176KB  (used by stack allocator right now)                   
        - StackSys     176KB  (obtained from system)                                
        - MSpanInuse   4KB    (mspan structures / in use now)                       
        - MSpanSys     176KB  (obtained from system)                                
        - MCacheInuse  1KB    (in use now))                                         
        - MCacheSys    16KB   (mcache structures / obtained from system)            
        - BuckHashSys  1MB    (profiling bucket hash table / obtained from system)  
        - GCSys        10MB   (GC metadata / obtained form system)                  
        - OtherSys     1MB    (other system allocations)                            

sleeping 30s; check memory usage with 'top -p 4129' now       1	30934787484 ns/op
ok  	sourcegraph.com/sourcegraph/appdash	30.982s
```

### top

```
$ top -p 4129

top - 13:45:03 up 2 days, 10:52,  9 users,  load average: 0.23, 0.37, 0.53
Tasks:   1 total,   0 running,   1 sleeping,   0 stopped,   0 zombie
%Cpu(s):  1.7 us,  1.5 sy,  0.0 ni, 96.7 id,  0.0 wa,  0.0 hi,  0.0 si,  0.0 st
KiB Mem:   8088044 total,  7269104 used,   818940 free,   194556 buffers
KiB Swap: 15625212 total,   975200 used, 14650012 free.  1373772 cached Mem

  PID USER      PR  NI    VIRT    RES    SHR S  %CPU %MEM     TIME+ COMMAND                          
 4129 stephen   20   0  270.6m  19.2m   5.7m S   2.6  0.2   0:01.18 appdash.test                     
```