@slimsag 2015-08-06
# Test 3

| Test Feature | Enabled |
|--------------|---------|
| 10 explicit runtime.GC calls? | Yes |
| 10 explicit runtime/debug.FreeOSMemory calls? | No |

```
$ go test -tags=danger -memprofilerate=1 -memprofile=mem.out -bench=BenchmarkChunkedCollector1mil -run=NONE -v
PASS
BenchmarkChunkedCollector1mil	

[pre] 1000000-collections:
  General statistics
    - Alloc       80KB   (allocated and still in use)                  
    - TotalAlloc  863KB  (allocated (even if freed))                   
    - Sys         2MB    (obtained from system (sum of XxxSys below))  
    - Lookups     5      (number of pointer lookups)                   
    - Mallocs     15973  (number of mallocs)                           
    - Frees       15446  (number of frees)                             

  Main allocation heap statistics
    - HeapAlloc     80KB   (allocated and still in use)         
    - HeapSys       832KB  (obtained from system)               
    - HeapIdle      456KB  (in idle spans)                      
    - HeapInuse     376KB  (in non-idle span)                   
    - HeapReleased  0B     (released to the OS)                 
    - HeapObjects   527    (total number of allocated objects)  

  Low-level fixed-size structure allocator statistics
    - StackInuse   192KB  (used by stack allocator right now)                   
    - StackSys     192KB  (obtained from system)                                
    - MSpanInuse   4KB    (mspan structures / in use now)                       
    - MSpanSys     16KB   (obtained from system)                                
    - MCacheInuse  1KB    (in use now))                                         
    - MCacheSys    16KB   (mcache structures / obtained from system)            
    - BuckHashSys  1MB    (profiling bucket hash table / obtained from system)  
    - GCSys        70KB   (GC metadata / obtained form system)                  
    - OtherSys     268KB  (other system allocations)                            

  

[post] 1000000-collections:
    General statistics
      - Alloc       214MB  (allocated and still in use)                  
      - TotalAlloc  356MB  (allocated (even if freed))                   
      - Sys         309MB  (obtained from system (sum of XxxSys below))  
      - Lookups     5      (number of pointer lookups)                   
      - Mallocs     76219  (number of mallocs)                           
      - Frees       34475  (number of frees)                             

    Main allocation heap statistics
      - HeapAlloc     214MB  (allocated and still in use)         
      - HeapSys       288MB  (obtained from system)               
      - HeapIdle      73MB   (in idle spans)                      
      - HeapInuse     214MB  (in non-idle span)                   
      - HeapReleased  0B     (released to the OS)                 
      - HeapObjects   41744  (total number of allocated objects)  

    Low-level fixed-size structure allocator statistics
      - StackInuse   192KB  (used by stack allocator right now)                   
      - StackSys     192KB  (obtained from system)                                
      - MSpanInuse   225KB  (mspan structures / in use now)                       
      - MSpanSys     240KB  (obtained from system)                                
      - MCacheInuse  1KB    (in use now))                                         
      - MCacheSys    16KB   (mcache structures / obtained from system)            
      - BuckHashSys  1MB    (profiling bucket hash table / obtained from system)  
      - GCSys        18MB   (GC metadata / obtained form system)                  
      - OtherSys     1MB    (other system allocations)                            

[garbage collections]
  0. GC - 62.225159ms
  1. GC - 56.665619ms
  2. GC - 89.450751ms
  3. GC - 65.649404ms
  4. GC - 57.154845ms
  5. GC - 58.463406ms
  6. GC - 56.052632ms
  7. GC - 56.512408ms
  8. GC - 57.36048ms
  9. GC - 75.21313ms

    

[after GC] 1000000-collections:
      General statistics
        - Alloc       128MB  (allocated and still in use)                  
        - TotalAlloc  356MB  (allocated (even if freed))                   
        - Sys         309MB  (obtained from system (sum of XxxSys below))  
        - Lookups     5      (number of pointer lookups)                   
        - Mallocs     77492  (number of mallocs)                           
        - Frees       72660  (number of frees)                             

      Main allocation heap statistics
        - HeapAlloc     128MB  (allocated and still in use)         
        - HeapSys       288MB  (obtained from system)               
        - HeapIdle      159MB  (in idle spans)                      
        - HeapInuse     128MB  (in non-idle span)                   
        - HeapReleased  0B     (released to the OS)                 
        - HeapObjects   4832   (total number of allocated objects)  

      Low-level fixed-size structure allocator statistics
        - StackInuse   192KB  (used by stack allocator right now)                   
        - StackSys     192KB  (obtained from system)                                
        - MSpanInuse   28KB   (mspan structures / in use now)                       
        - MSpanSys     240KB  (obtained from system)                                
        - MCacheInuse  1KB    (in use now))                                         
        - MCacheSys    16KB   (mcache structures / obtained from system)            
        - BuckHashSys  1MB    (profiling bucket hash table / obtained from system)  
        - GCSys        18MB   (GC metadata / obtained form system)                  
        - OtherSys     1MB    (other system allocations)                            

sleeping 30s; check memory usage with 'top -p 4362' now       1	31641170518 ns/op
ok  	sourcegraph.com/sourcegraph/appdash	31.684s
```

### top

```
$ top -p 4362

top - 13:55:41 up 2 days, 11:02,  9 users,  load average: 0.24, 0.33, 0.44
Tasks:   1 total,   0 running,   1 sleeping,   0 stopped,   0 zombie
%Cpu(s):  1.6 us,  0.8 sy,  0.0 ni, 97.7 id,  0.0 wa,  0.0 hi,  0.0 si,  0.0 st
KiB Mem:   8088044 total,  7761588 used,   326456 free,   195860 buffers
KiB Swap: 15625212 total,   974988 used, 14650224 free.  1395596 cached Mem

  PID USER      PR  NI    VIRT    RES    SHR S  %CPU %MEM     TIME+ COMMAND                          
 4362 stephen   20   0  473.3m 314.9m   5.7m S   2.1  4.0   0:01.98 appdash.test                     
```